code/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
factory_boy>=3.2.0
pytest-django>=4.4.0
pytest-order>=1.0.0
